+++ src/rpc/rpc_util.rs
@@ -2,7 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use crate::rpc_api::{
-    auth_api::*, check_access, data_types::JsonRpcServerState, eth_api::ETH_BLOCK_NUMBER,
+    auth_api::*,
+    check_access,
+    data_types::JsonRpcServerState,
+    eth_api::{ETH_BLOCK_NUMBER, ETH_CHAIN_ID},
     ACCESS_MAP,
 };
 use http::{HeaderMap, HeaderValue, StatusCode};
@@ -42,7 +45,7 @@ pub fn is_streaming_method(method_name: &str) -> bool {
     STREAMING_METHODS.contains(&method_name)
 }
 
-const V1_METHODS: [&str; 1] = [ETH_BLOCK_NUMBER];
+const V1_METHODS: [&str; 2] = [ETH_BLOCK_NUMBER, ETH_CHAIN_ID];
 
 pub fn is_v1_method(method_name: &str) -> bool {
     V1_METHODS.contains(&method_name)