+++ src/rpc_api/mod.rs
@@ -95,6 +95,7 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(state_api::STATE_GET_RANDOMNESS_FROM_TICKETS, Access::Read);
     access.insert(state_api::STATE_GET_RANDOMNESS_FROM_BEACON, Access::Read);
     access.insert(state_api::STATE_READ_STATE, Access::Read);
+    access.insert(state_api::STATE_CIRCULATING_SUPPLY, Access::Read);
     access.insert(state_api::STATE_SECTOR_GET_INFO, Access::Read);
     access.insert(
         state_api::STATE_VM_CIRCULATING_SUPPLY_INTERNAL,
@@ -123,6 +124,7 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(node_api::NODE_STATUS, Access::Read);
 
     // Eth API
+    access.insert(eth_api::ETH_CHAIN_ID, Access::Read);
     access.insert(eth_api::ETH_BLOCK_NUMBER, Access::Read);
 
     access
@@ -366,4 +368,5 @@ pub mod node_api {
 // Eth API
 pub mod eth_api {
     pub const ETH_BLOCK_NUMBER: &str = "Filecoin.EthBlockNumber";
+    pub const ETH_CHAIN_ID: &str = "Filecoin.EthChainId";
 }